ESASKYPCR-1648 - SizeFormatter now accepts integers instead of long
git-svn-id: https://satscm.esac.esa.int/esasky/esasky-app/trunk@4431 aa7853a0-da02-4334-bf23-07ffea9f4f01

+++ esasky-cl/src/main/java/esac/archive/esasky/cl/web/client/utility/SizeFormatter.java
@@ -2,7 +2,7 @@ package esac.archive.esasky.cl.web.client.utility;
 
 public final class SizeFormatter {
 
-	public final native static String formatBytes(Long bytes, int decimals) /*-{ 
+	public final native static String formatBytes(int bytes, int decimals) /*-{ 
 	   if(bytes == 0) return '0 Bytes';
 	   var k = 1024, 
 	   dm = decimals <= 0 ? 0 : decimals || 2,
